Abstract buffer resource to separate class to share with upstream
